Вынесла кнопку на страницу CatalogPage

+++ MySupperShop/Pages/ProductPage.razor.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Components;
+using MudBlazor;
 using MySupperShop.Models;
 using MySupperShopHttpApiClient.Interfaces;
 
@@ -12,38 +13,46 @@ namespace MySupperShop.Pages
         public IMyShopClient? ShopClient { get; set; }
         [Inject]
         public NavigationManager manager { get; set; }
-
-        public string ProductFieldDeleted { get; set; } = string.Empty;
+        [Inject]
+        public IDialogService DialogService { get; set; }
+        string state = string.Empty;
         private Product? _product;
         private CancellationTokenSource _cts = new CancellationTokenSource();
 
         protected override async Task OnInitializedAsync()
         {
-                _product = await ShopClient!.GetProduct(ProductId, _cts.Token);
+            _product = await ShopClient!.GetProduct(ProductId, _cts.Token);
         }
         public async Task DeleteProduct()
         {
             try
             {
-                await ShopClient!.DeleteProduct(_product!, _cts.Token);
-                ProductFieldDeleted = "Товар удален!";
-                await InvokeAsync(() => StateHasChanged());
-                await Task.Delay(TimeSpan.FromSeconds(5), _cts.Token);
-                manager.NavigateTo("/catalog");
+                bool? result = await DialogService.ShowMessageBox(
+            "Информация",
+            "Вы верены, что хотите удалить продукт?",
+            yesText: "Да", cancelText: "Нет");
+                state = result == null ? "No" : "Yes";
+                if (state == "Yes")
+                {
+                    await ShopClient!.DeleteProduct(_product!, _cts.Token);
+                    await DialogService.ShowMessageBox("Информация", "Товар удален!");
+                    await Task.Delay(TimeSpan.FromSeconds(5), _cts.Token);
+                    manager.NavigateTo("/catalog");
+                }
+                if (state == "No")
+                {
+                    await DialogService.ShowMessageBox("Информация", "Товар не удален!");
+                    return;
+                }
             }
             catch (ArgumentNullException)
             {
-                ProductFieldDeleted = "Товар не удален!";
-                await InvokeAsync(() => StateHasChanged());
+                await DialogService.ShowMessageBox("Ошибка", "Товар не удален!");
             }
         }
         public void ToProductEditPage()
         {
             manager.NavigateTo($"/products/{_product!.Id}/editor");
-        }
-        public void ToAddProductPage()
-        {
-            manager.NavigateTo($"/products/new");
         }
         public void Dispose()
         {
